requirements.txt
fastmcp
langchain
langchain-openai
langchain-mcp-adapters
python-dotenv